Allow building multiple entrypoints
You can now pass --bin to `cpkg run` or `cpkg build` to build your cpkg project using a certain entrypoint.

This works by telling the linker to allow duplicate definitions in these cases, and first compiling that file.

Traditional `cpkg run` or `cpkg build` will not be affected by this.

(There's also a message given on "multiple definition of main" errors hinting to use --bin).

If you choose to build a specific entrypoint, it will export with that entrypoint as the filename in target. So `cpkg build --bin main2` will build to `target/main2`.